build: 빌드 에러 해결 #14
