+++ src/app/(pages)/(club)/club/_components/oneTimeClub/DateTime.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 import { useState } from "react";
